requirements.txt
streamlit==1.42.0
pandas==2.2.3
python-dotenv==1.0.1
selenium==4.28.1